% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_codigo_diputados.R
\name{asignar_diputados}
\alias{asignar_diputados}
\alias{calcular_nd1}
\alias{calcular_nd2}
\alias{corregir_nd2_sobrerep}
\alias{calcular_nd2_1}
\title{Functions that help with the  assignment of the Chamber of Deputies.}
\usage{
calcular_nd1(theta_long, coa_long, keys_part_coa)

calcular_nd2_1(theta_long, lista_nominal_estrato, keys_part_coa)

corregir_nd2_sobrerep(eta_long, nd1_long, nd2_1_long)
}
\arguments{
\item{theta_long}{a data frame in long format with columns:
\code{id_estrato} (integer for each of the 300 federal districts);
\code{id_partido} (integer for each of the parties);
\code{sim} (integer id for each of the simulations);
\code{value} (numeric with the value of the estimated theta)}

\item{coa_long}{tibble in long format with the data for coalition assignment in each Federal District
\code{id_estrato} (integer for each of the 300 federal districts);
\code{coaliciones_que_aplican} (character, with coalitions that apply in the DF,
either name of coalition to match \code{coalicion_partido} in \code{keys_part_coa}),
"Ambas" if both/all or "Ninguna" if no coalitions apply for that DF;
\code{nombre_partido} (character with party name);
\code{reciben_votos_coa} (integer, indicator: 1 if the party of the coalition receives the votes for that DF, 0 if not);}

\item{keys_part_coa}{a data frame with the key variables for assignment.
\code{id_partido} (integer for each of the parties);
\code{nombre_partido} (character with party name);
\code{candidato_independiente} (numeric, 1 if the party is from an independent candidate, 0 if not);
\code{id_coalicion} (numeric, id for coalicion);
\code{coalicion_partido} (character, name of the coalicion);}

\item{lista_nominal_estrato}{a data frame with the nominal list for each DF. Columns:
\code{id_estrato} (integer id for each of the 300 federal districts);
\code{LISTA_NOMINAL} (integer, nominal list for each of the 300 federal districts);}

\item{eta_long}{tibble in long format with with \code{sim}, \code{id_partido}, \code{nombre_partido} and
\code{eta} which is created in \code{calcular_nd2} and used in \code{corregir_nd2_sobrerep} as an input.}

\item{nd1_long}{tibble in long format with grouped by simulation with columns:
\code{sim}, \code{id_partido}, \code{nombre_partido}, and \code{nd1}.
\code{nd1_long} is created in \code{calcular_nd1} and used in \code{corregir_nd2_sobrerep} as an input}

\item{nd2_1_long}{tibble in long format with grouped by simulation with columns:
\code{sim}, \code{id_partido}, \code{nombre_partido}, and \code{nd2_1}.
\code{nd2_long} is created in \code{calcular_nd2} and used in \code{corregir_nd2_sobrerep} as an input}
}
\value{
\code{calcular_nd1} returns a tibble (\code{nd1_long}) grouped by simulation with columns:
\code{sim}, \code{id_partido}, \code{nombre_partido}, and
\code{nd1} (integer, number of seats out of the first 300 assigned for each of the parties);

\code{calcular_nd2} returns a list with 2 elements:
\code{nd2_1_long} is a tibble grouped by simulation with columns:
\code{sim}, \code{id_partido}, \code{nombre_partido} and
\code{nd2_1} (numeric, number of seats out of the first assignment of the 200 seats assigned for each of the parties);
and \code{eta_long} another tibble with \code{sim}, \code{id_partido}, \code{nombre_partido} and \code{eta} which is used in the next function for corrections.

\code{corregir_nd2_sobrerep} uses outputs from the other two functions and returns a tibble with columns:
\code{sim}, \code{id_partido}, \code{nombre_partido} and
\code{nd2} (numeric, number of seats out of the corrections for over representation of the 200 seats assigned for each of the parties);
}
\description{
There are 500 seats to be assigned in the Chamber of Deputies in Mexico.
The process involves different steps.
\code{calcular_nd1}: Calculates the assignment of the first 300 seats by "maximo relativo".
\code{calcular_nd2}: Calculates the assignment of the second 200 seats by "representacion proporcional".
\code{corregir_nd2_sobrerep}: Corrects the second assignment of the second 200 seats for over representation,
taking into account the maximum number allowed per party.
}
