% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_estimation.R
\name{assign_majority}
\alias{assign_majority}
\title{Assign majority party to each stratum and bootstrap repetition}
\usage{
assign_majority(
  estimates_strata_tbl,
  coalitions_tbl,
  party_name,
  candidate_name
)
}
\arguments{
\item{estimates_strata_tbl}{Tibble with estimates of proportion of votes per party per repetition (output of bootstrap_diputados)}

\item{coalitions_tbl}{Tibble with party names and corresponding assigned party by stratum.}

\item{party_name}{Unquoted variable indicating the party name.}

\item{candidate_name}{Unquoted variable indicating the assigned party.}
}
\description{
Assign majority party to each stratum and bootstrap repetition
}
