% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conteo_2018.R
\docType{data}
\name{conteo_2018}
\alias{conteo_2018}
\title{Computos distritales for 2018 electino}
\format{
A data frame with 156840 rows and 95 variables:
\describe{
\item{CLAVE_CASILLA - JAMK}{computos distritales 2018}
\item{TVIVPARHAB-VPH_SNBIEN}{2010 INEGI census data}
}
}
\source{
\url{https://computos2018.ine.mx/#/presidencia/nacional/1/1/1/1}
}
\usage{
conteo_2018
}
\description{
Contains results of vote counts for the 2018 presidential election in Mexico
}
\keyword{datasets}
