% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_estimation.R
\name{hb_estimation_parallel}
\alias{hb_estimation_parallel}
\title{Bayesian hierarchical model to estimate proportion of votes allocated to each party}
\usage{
hb_estimation_parallel(
  data_tbl,
  sampling_frame,
  split_var = NULL,
  nominal_list_var = NULL,
  num_cores = 5,
  inv_metric_list = NULL,
  nominal_max = 1000,
  frac,
  sig_figs = 8,
  ...
)
}
\arguments{
\item{data_tbl}{\code{tibble} of sample data.}

\item{sampling_frame}{\code{tibble} of sampling frame with stratum variable (named exactly as in
\code{data_tbl}) and covariates.}

\item{split_var}{unquoted name of variable that splits sampling frame and data}

\item{nominal_list_var}{Unquoted name of variable with nominal list of voters}

\item{num_cores}{Number of cores to use for parallel computation}

\item{inv_metric_list}{List of inv_metric diagonals guesses for each split}

\item{nominal_max}{Maximum number of nominal count for stations. Used for
stations without fixed nominal list.}

\item{frac}{Sampling fraction for full sample size.}

\item{sig_figs}{Number of significant figures for Stan output}

\item{...}{Other parameters passed to hb_estimation.}
}
\value{
A list with model fit (if return_fit=TRUE), a \code{tibble}
estimates including point estimates for each party (median)
and limits of credible intervals, and a vector inv_metric for the model
}
\description{
Compute point estimate and credible intervals for each candidate, parallized over
split_var.
}
\details{
Posterior simulations of parameters are computed using stan,
and each party's votes are simulated for every polling station (logit model)
or with a softmax link for the default (mlogit model). There is one
independent model for each party, and proportions are calculated from posterior simulations
of total votes. Splits are modelled independently
}
