% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marco_2018.R
\docType{data}
\name{marco_nal_2018}
\alias{marco_nal_2018}
\title{Sampling frame for elections 2018.}
\format{A data frames:}
\source{
\url{https://cartografia.ife.org.mx}
}
\usage{
marco_nal_2018
}
\description{
Datasets containing the 2018 sampling frame for Mexican national election, 
and sampling frames for Governor elections in Chiapas, Guanajuato, and 
Morelos. Each row corresponds to a polling station and the variables include 
strata, nominal list, and other descriptorss of the polling station.
}
\keyword{datasets}
