% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mich_2015.R
\docType{data}
\name{mich_2015}
\alias{mich_2015}
\title{Computos distritales for 2015 state election}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5790 rows and 94 columns.
}
\source{
\url{http://computos2015.ine.mx}
}
\usage{
mich_2015
}
\description{
Contains results of vote counts for the 2015 state elections in Mexico
for Michoacán
}
\keyword{datasets}
