% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrp_party_estimation.R
\name{mrp_party_estimation}
\alias{mrp_party_estimation}
\title{Postratified estimation for the number of votes for a given candidate.}
\usage{
mrp_party_estimation(data, party, stratum, frac_sample = 1, n_chains = 3,
  n_iter = 2000, n_burnin = 500, seed = NA, model_string = NA)
}
\arguments{
\item{data}{A \code{data.frame} with variables: ln_total, region,
distrito_loc_17, tamano_md, tamano_gd, casilla_ex and the column with
number of votes for the party.}

\item{party}{unquoted variable indicating the column from the data.frame to
be modeled.}

\item{stratum}{If sampling the data, unquoted variable indicating the column
from the data.frame to be used as strata. The strata will also be used in
the hierarchical structure of the model.}

\item{frac_sample}{If sampling the data, numeric value indicating the fraction
of the data to sample, the sample is selected using stratified sampling
with probability proportional to size.}

\item{n_iter, n_chains, n_burnin}{Number of iterations, chains and burnin size
to be used in \code{\link[R2jags]{jags}}.}

\item{seed}{Integer value used to set the state of the random number generator.}

\item{modelo_jags}{optional string specifying variations of the jags model.}
}
\value{
A \code{list} with the object fitted using R2jags::jags and the vector
  of simulated counts per candidate.
}
\description{
The function fits a model using the \code{\link[R2jags]{jags}} function and
predicts number of votes for a given candidate in unobserved polling
stations.
}
\details{
The default model is:
  \deqn{X_k \sim N(n_k \theta_k, n_{k}\sigma^2)}
  \deqn{
  \theta_k=\beta^0 + \beta^{rural}\cdot rural_k + \beta^{tipoSp}\cdot
   tipoSp_k + \\ \beta^{tipoEx}\cdot tipoEx_k + \beta^{tamanoMd}\cdot
   tamanoMd_k + \beta^{tamanoGd}\cdot tamanoGd_k +
   \beta^{distrito}_{distrito(k)} }
   \deqn{\beta_{dl}\sim N(\beta_{region(k)}^{region}, \sigma_{dl}^2)}
}
\examples{
data("gto_2012")
mrp_party_estimation(gto_2012, party = pan_na, stratum = distrito_loc_17,
  frac_sample = 1, seed = 2212)
}
