% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrp_party_estimation_stan.R
\name{mrp_party_estimation_stan}
\alias{mrp_party_estimation_stan}
\title{Postratified estimation for the number of votes for a given candidate.}
\usage{
mrp_party_estimation_stan(data, party, stratum, frac = 1, n_iter = 300,
  warmup = 150, n_chains = 2, seed = NA, model_string = NULL,
  set_strata_na = integer(0))
}
\arguments{
\item{data}{A \code{data.frame} with variables: ln_total, region,
distrito_loc_17, tamano_md, tamano_gd, casilla_ex and the column with
number of votes for the party.}

\item{party}{Unquoted variable indicating the column from the data.frame to
be modeled.}

\item{stratum}{If sampling the data, unquoted variable indicating the column
from the data.frame to be used as strata. The strata will also be used in
the hierarchical structure of the model.}

\item{frac}{If sampling the data, numeric value indicating the fraction
of the data to sample, the sample is selected using stratified sampling
with probability proportional to size.}

\item{n_iter, warmup}{Number of iterations, number of warmup iterations.
to be used in \code{\link[rstan]{sampling}}.}

\item{seed}{Integer value used to set the state of the random number
generator.}

\item{model_string}{String indicating the model to be used, defaults to
\code{"neg_binomial"}.}
}
\value{
A list containing fit object and vector
  the simulated counts for the candidate.
}
\description{
The function fits a model using the \code{rstan} package and
predicts number of votes for a given candidate in unobserved polling
stations. Optionally the model can be fit with a stratified random sample
of the data, and predict for the population.
}
\examples{
mrp_party <- mrp_party_estimation_stan(nal_2012, pri_pvem, frac = 0.02,
    stratum = estrato, n_iter = 300, warmup = 150,
    n_chains = 2, seed = 19291)
quantile(mrp_party$y)
}
