% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nal_2006.R
\docType{data}
\name{nal_2006}
\alias{nal_2006}
\title{Presidential electoral results 2006.}
\format{A data frames:}
\source{
\url{https://cartografia.ife.org.mx}
}
\usage{
nal_2006
}
\description{
A dataset containing the election results for president 2006, each row
corresponds to a polling station. The data provided by INE lacked some 
covariates corresponding to 2006, when this occurs we use the corresponding
value in 2017.
}
\keyword{datasets}
