% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnp_estimation.R
\name{nnp_estimation}
\alias{nnp_estimation}
\title{Normal No-Pooling model}
\usage{
nnp_estimation(data, ln, stratum, data_stratum, n_stratum, ..., n_sims = 1000)
}
\arguments{
\item{data}{\code{data.frame}}

\item{ln}{Unquoted variable indicating the nominal list (number of potential
voters) at each polling station.}

\item{stratum}{Unquoted variable indicating the stratum for each polling
station.}

\item{data_stratum}{Data frame with stratum variable (named exactly as in
\code{data}) and number of polling stations per strata.}

\item{n_stratum}{Unquoted variable indicating the number of polling stations
in each stratum.}

\item{...}{Unquoted variables indicating the number of votes in each polling
station for each candidate.}

\item{std_errors}{Logical value indicating whether to compute standard errors
(using bootstrap), defaults to TRUE.}

\item{B}{Number of bootstrap replicates used to compute standard errors,
defaults to 50.}

\item{seed}{integer value used to set the state of the random number
generator (optional). It will only be used when computing standard errors.}
}
\value{
A list with two entries:
\enumerate{
\item lambdas_summary \code{data.frame} including posterior mean, median,
standard error, and quantiles (0.025 and 0.975) for each party.
\item lamdas_sim \code{data.frame} with simulations for each party.
}
}
\description{
This function implements the Normal No-Pooling model from \cite{AA}.
}
\examples{
# count number of polling stations per stratum

zac_stratum_sizes <- zac_2015 \%>\%
    dplyr::group_by(ID_DISTRITO_15) \%>\%
    dplyr::summarise(n_stratum = n())
zac_sample <- select_sample_prop(zac_2015, stratum = ID_DISTRITO_15, 0.06) \%>\%
  filter(LISTA_NOMINAL_15 > 0)
zac_nnp <- nnp_estimation(data = zac_sample, ln = LISTA_NOMINAL_15,
  stratum = ID_DISTRITO_15,
  data_stratum = zac_stratum_sizes, n_stratum = n_stratum,
  contains(CAND), n_sims = 100)
zac_nnp$lambdas_summary
}
