% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_predictive_check.R
\name{pp_check}
\alias{pp_check}
\alias{pp_check_party}
\title{Posterior predictive checks for bayesian vote count models.}
\usage{
pp_check_party(model_fit_party)

pp_check(model_fit)
}
\arguments{
\item{model_fit_party}{Output of \code{mrp_party_estimation}}

\item{model_fit}{Output of \code{mrp_estimation}}
}
\value{
A \code{tbl} with posterior simulations for observed
polling stations, with observed vote counts
}
\description{
Posterior predictive checks for bayesian vote count models.
}
\examples{
data("gto_2012")
model <- mrp_party_estimation(gto_2012, party = pan_na, stratum = distrito_loc_17,
  frac = 1, seed = 2212)
pp_check <- pp_check_party(model)
}
