% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickcountmx-package.R
\docType{package}
\name{quickcountmx-package}
\alias{quickcountmx}
\alias{quickcountmx-package}
\title{quickcountmx: Quick-count estimation package for the 2021 Mexican election}
\description{
Functions to simulate and estimate Mexican election results based on a simple or stratified random sample, the functions were used, among other methodologies, to anticipate the final results of the 2021-2022 Mexican elections.
}
\author{
\strong{Maintainer}: Teresa Ortiz \email{teresa.ortiz.mancera@gmail.com}

Authors:
\itemize{
  \item Felipe Gonzalez \email{felipexgonzalez@gmail.com}
}

}
\keyword{internal}
