% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickcountmx.R
\docType{package}
\name{quickcountmx}
\alias{quickcountmx}
\title{quickcountmx}
\description{
Functions to simulate and estimate Mexican election results based on a simple or stratified random sample, the functions were used, among other methodologies, to anticipate the final results of the 2021 Mexican elections.
}
