% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regiones.R
\docType{data}
\name{regiones}
\alias{regiones}
\title{Regiones}
\format{A data frame with 32 rows and 2 variables:
\describe{
  \item{area}{geographical area.}
  \item{id_estado}{State id}
}}
\source{
\url{https://cartografia.ife.org.mx}
}
\usage{
regiones
}
\description{
A dataset containing a stratification of Mexico's states, this used to 
speed up estimations for the national model.
}
\keyword{datasets}
