% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_arrivals.R
\name{simulate_arrivals}
\alias{simulate_arrivals}
\title{\code{simulate_arrivals} Simulate arrival times based on parametric bootstrap
of \code{model}}
\usage{
simulate_arrivals(id, model, new_data_tbl = NULL, hour_censoring = 5)
}
\arguments{
\item{id}{id for the simulation run}

\item{model}{fit with function \code{fit_model}}

\item{new_data_tbl}{new data to simulate times (if not for the original dataset
for which the model was fit)}

\item{hour_censoring}{add censoring structure if desired certain hours after
polling station closing (use \code{Inf} for the exact simulated times )}
}
\description{
\code{simulate_arrivals} Simulate arrival times based on parametric bootstrap
of \code{model}
}
