% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_estimation.R
\name{assign_deputy_seats}
\alias{assign_deputy_seats}
\alias{assign_majority}
\alias{assign_all_seats}
\alias{assign_prop}
\title{Functions that help with the  assignment of the Chamber of Deputies.}
\usage{
assign_majority(
  estimates_strata_tbl,
  assignment_tbl,
  party_name,
  candidate_name
)

assign_all_seats(reps_list, assignment_tbl)

assign_prop(total_tbl)
}
\arguments{
\item{estimates_strata_tbl}{Tibble with estimates of proportion of votes per party per repetition per stratum (output of bootstrap_diputados)}

\item{assignment_tbl}{Tibble with party names and corresponding assigned party by stratum.}

\item{party_name}{Unquoted variable indicating the party name.}

\item{candidate_name}{Unquoted variable indicating the assigned party.}

\item{reps_list}{A named list (output of bootstrap_diputados) with elements strata_tbl and total_tbl}

\item{total_tbl}{Tibble with estimates of proportion of votes per party per repetition (output of bootstrap_diputados)
total_tbl must include columns: n_assign seats to allocate, per replication and
topped: a logical value whether the party has reached it's maximum number of seats (TRUE)}
}
\value{
\code{assign_majority} returns a tibble with columns:
\code{rep} number of repetition, \code{candidate} the party, and
\code{n_seats_maj} integer, number of seats assigned by majority for each of the parties out of the first 300;

\code{assign_all_seats} returns a tibble with columns:
\code{rep} number of repetition, \code{candidate} the party, and
\code{n_seats_maj} integer, number of seats assigned by majority for each of the parties out of the first 300;
\code{n_seats_prop} integer, is the number of seats assigned by proportionality and "resto mayor", out of the 200, for each of the parties;
\code{n_seats_total} integer, is the total number of seats, out of the 500, for each of the parties;

\code{assign_prop} returns a tibble where :
\code{rep} is the number of repetition, \code{party} is the name of the party, and
\code{n_seats_prop} integer, is the number of seats assigned by proportionality and "resto mayor" for each of the parties;
}
\description{
There are 500 seats to be assigned in the Chamber of Deputies in Mexico.
The process involves different steps.
\code{assign_all_seats}: Assigns the total of 500 seats of majority and proportional seats, checking for maximum constraints per
\code{assign_majority}: Add column defining the majority party of the first 300 seats to each stratum and repetition.
\code{assign_prop}: Assigns n_assign seats to parties using proportional representation in national voting and "resto mayor".
}
