% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_estimation.R
\name{bootstrap_diputados}
\alias{bootstrap_diputados}
\title{Bootstrap replicates to estimate diputados results}
\usage{
bootstrap_diputados(
  data_tbl,
  stratum,
  stratum_tbl,
  n_stratum,
  coalitions_tbl,
  B = 50,
  seed = NA,
  samples_table = FALSE
)
}
\arguments{
\item{data_tbl}{\code{tibble}}

\item{stratum}{Unquoted variable indicating the stratum for each polling
station.}

\item{n_stratum}{Unquoted variable indicating the number of polling stations
in each stratum.}

\item{coalitions_tbl}{Tibble with coalitions names and corresponding parties.}

\item{B}{Number of bootstrap replicates,
defaults to 50.}

\item{seed}{integer value used to set the state of the random number
generator (optional).}

\item{samples_table}{logical indicating if the function should return the samples
as a list of tibbles for bootstrap samples}

\item{data_stratum}{Data frame with stratum variable (named exactly as in
\code{data}) and number of polling stations per strata.}
}
\value{
A \code{list} including two componentes: point estimates at national level
and at stratum level, and bootstrap replications of these quantities.
}
\description{
Compute bootstrap resamples of ratio estimator for each party at national level,
along with proportion of votes for each stratum and party.
}
