% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_estimation.R
\name{bootstrap_estimation_diputados}
\alias{bootstrap_estimation_diputados}
\title{Bootstrap estimates for diputados}
\usage{
bootstrap_estimation_diputados(
  data_tbl,
  stratum,
  stratum_tbl,
  n_stratum,
  coalitions_tbl,
  assignment_tbl,
  B = 50,
  seed = NA,
  samples_table = FALSE
)
}
\description{
Compute bootstrap confidence intervals using ratio estimator for each party at national level,
along with proportion of votes for each stratum and party.
}
