% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration}
\alias{calibration}
\alias{calibration_party}
\alias{calibration_prop}
\alias{summary_calibration_party}
\alias{summary_calibration}
\title{Calibration}
\usage{
calibration_party(data, party, stratum, frac = 1, n_iter = 2000,
  n_burnin = 500, n_chains = 3, seed = NA, cl_cores = 14, n_rep = 5,
  model_string = NULL)

calibration_prop(data, ..., stratum, frac = 1, n_iter = 2000,
  n_burnin = 500, n_chains = 3, seed = NA, cl_cores = 3, n_rep = 5,
  model_string = NULL, num_missing_strata = 0)

summary_calibration_party(calib_run_party, alpha_r = 0.05)

summary_calibration(calib_run, alpha_r = 0.05)
}
\arguments{
\item{data}{A \code{data.frame} with variables: ln_total, region,
distrito_loc_17, tamano_md, tamano_gd, casilla_ex and the column with
number of votes for the party.}

\item{party}{Unquoted variable indicating the column from the data.frame to
be modeled.}

\item{stratum}{If sampling the data, unquoted variable indicating the column
from the data.frame to be used as strata. The strata will also be used in
the hierarchical structure of the model.}

\item{frac}{If sampling the data, numeric value indicating the fraction
of the data to sample, the sample is selected using stratified sampling
with probability proportional to size.}

\item{n_iter}{Number of iterations, burnin size and chains.
to be used in \code{\link[R2jags]{jags}}.}

\item{n_burnin}{Number of iterations, burnin size and chains.
to be used in \code{\link[R2jags]{jags}}.}

\item{n_chains}{Number of iterations, burnin size and chains.
to be used in \code{\link[R2jags]{jags}}.}

\item{seed}{An integer vector of length 7 to be send to
\code{\link[parallel]{clusterSetRNGStream}}.}

\item{cl_cores}{Number of cores, parameter is used in
\code{\link[parallel]{makeCluster}}.}

\item{n_rep}{Number of repetitions of sample selection and model fitting.}

\item{model_string}{String indicating the model to fit.}

\item{alpha_r}{Calibration examines coverage of (1-alpha_r)*100 intervals.}
}
\value{
\code{data.frame} with posterior simulation of total votes (if using
  \code{calibration_party}) or proportions (if using \code{calibration_prop})
  for each repetiton of sample selection and model fitting.
}
\description{
The functions \code{calibration_party} and \code{calibration_prop} perform 
simulation from past election results, fit a model calling
(\code{\link{mrp_party_estimation}} or \code{\link{mrp_estimation}}) for each 
simulation and return posterior simulations along actual outcomes, these can 
later be used to compute calibration summaries with the corresponding 
function: \code{summary_calibration_party} or \code{summary_calibration}. 
\code{calibration_party} is useful to calibrate individual models of total 
votes for a given party and \code{calibration_prop} is useful to analyse 
estimates for proportion of votes per party.
}
\details{
The functions are computationally demanding, they were designed
to run on computers with more than 16 cores.
}
\seealso{
\code{\link{mrp_estimation}}, \code{\link{mrp_party_estimation}}
}
