% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conteo_2018.R
\docType{data}
\name{conteo_2018}
\alias{conteo_2018}
\title{Computos distritales for 2018 election}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 156432 rows and 133 columns.
}
\source{
\url{https://computos2018.ine.mx/#/presidencia/nacional/1/1/1/1}
}
\usage{
conteo_2018
}
\description{
Contains results of vote counts for the 2018 presidential election in Mexico
}
\keyword{datasets}
