% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_estimation.R
\name{hb_estimation}
\alias{hb_estimation}
\title{Bayesian hierarchical model to estimate proportion of votes allocated to each party}
\usage{
hb_estimation(
  data_tbl,
  stratum,
  id_station,
  sampling_frame,
  parties,
  covariates,
  prop_obs = 0.995,
  seed = NULL,
  return_fit = FALSE,
  num_iter = 200,
  num_warmup = 200,
  adapt_delta = 0.8,
  max_treedepth = 10,
  chains = 3,
  sig_figs = 6,
  model = "mlogit-corr",
  nominal_max = 1200,
  threads_per_chain = 1,
  inv_metric = NULL,
  erase_output_files = TRUE
)
}
\arguments{
\item{data_tbl}{\code{tibble} of sample data.}

\item{stratum}{Unquoted variable indicating the stratum for each polling
station.}

\item{id_station}{Unquoted variable indicating the id for each polling
station.}

\item{sampling_frame}{\code{tibble} of sampling frame with stratum variable (named exactly as in
\code{data_tbl}) and covariates.}

\item{parties}{Unquoted variables indicating the number of votes in each polling
station for each candidate.}

\item{covariates}{Unquoted variables indicating the covariates in each polling
station.}

\item{prop_obs}{Proportion of size of observed sample to total designed sample.}

\item{seed}{integer value used to set the state of the random number
generator (optional).}

\item{return_fit}{Returns summary if FALSE (default), otherwise return cmdstanr fit}

\item{num_iter}{Number of post warmup iterations}

\item{num_warmup}{Number of warmup iterations}

\item{adapt_delta}{The adaptation target acceptance statistic (default 0.80.}

\item{max_treedepth}{The maximum allowed tree depth for the NUTS engine (default 10)}

\item{chains}{Number of chains (will be run in parallel)}

\item{sig_figs}{Number of significant figures for Stan output}

\item{model}{One of "mlogit" (the default) or "logit"}

\item{nominal_max}{Maximum number of nominal count for stations. Used for
stations without fixed nominal list.}

\item{threads_per_chain}{Number of threads per chain to split calculation of log-posterior}

\item{inv_metric}{vector of inverse metric diagonal for the model. Default is NULL}

\item{erase_output_files}{If TRUE, output files are erased after fitting}
}
\value{
A list with model fit (if return_fit=TRUE), a \code{tibble}
estimates including point estimates for each party (median)
and limits of credible intervals, and a vector inv_metric for the model
}
\description{
Compute point estimate and credible intervals for each candidate.
}
\details{
Posterior simulations of parameters are computed using stan,
and each party's votes are simulated for every polling station (logit model)
or with a softmax link for the default (mlogit model). There is one
independent model for each party, and proportions are calculated from posterior simulations
of total votes.
}
