% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_estimation.R
\name{hb_estimation}
\alias{hb_estimation}
\title{Bayesian hierarchical model to estimate proportion of votes allocated to each party}
\usage{
hb_estimation(
  data_tbl,
  stratum,
  id_station,
  data_stratum,
  parties,
  covariates,
  prop_obs = 0.995,
  seed = NA,
  return_fit = FALSE,
  num_iter = 1000,
  chains = 3
)
}
\arguments{
\item{data_tbl}{\code{tibble} of sample data}

\item{stratum}{Unquoted variable indicating the stratum for each polling
station.}

\item{id_station}{Unquoted variable indicating the idfor each polling
station.}

\item{data_stratum}{Data frame with stratum variable (named exactly as in
\code{data}) and number of polling stations per strata.}

\item{parties}{Unquoted variables indicating the number of votes in each polling
station for each candidate.}

\item{covariates}{Unquoted variables indicating the covariates in each polling
station.}

\item{prop_obs}{Proportion of size of observed sample to total designed sample.}

\item{seed}{integer value used to set the state of the random number
generator (optional).}

\item{return_fit}{Returns summary if FALSE (default), otherwise return cmdstanr fit}

\item{num_iter}{Number of post warmup iterations}

\item{chains}{Number of chains (will be run in parallel)}
}
\value{
A \code{tibble} including point estimates for each party (median)
and limits of credible intervals.
}
\description{
Compute point estimate and credible intervals for each candidate.
}
\details{
Posterior simulations of parameters are computed using stan,
and each party's votes are simulated en every polling station. There is one
independent model for each party, and proportions are calculated from posterior simulations
of total votes.
}
