% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_estimation.R
\name{hb_estimation}
\alias{hb_estimation}
\title{Bayesian hierarchical model to estimate proportion of votes allocated to each party}
\usage{
hb_estimation(
  data_tbl,
  stratum,
  id_station,
  sampling_frame,
  parties,
  covariates,
  prop_obs = 0.995,
  seed = NA,
  return_fit = FALSE,
  num_iter = 1000,
  chains = 3,
  model = "mlogit"
)
}
\arguments{
\item{data_tbl}{\code{tibble} of sample data.}

\item{stratum}{Unquoted variable indicating the stratum for each polling
station.}

\item{id_station}{Unquoted variable indicating the id for each polling
station.}

\item{sampling_frame}{\code{tibble} of sampling frame with stratum variable (named exactly as in
\code{data_tbl}) and covariates.}

\item{parties}{Unquoted variables indicating the number of votes in each polling
station for each candidate.}

\item{covariates}{Unquoted variables indicating the covariates in each polling
station.}

\item{prop_obs}{Proportion of size of observed sample to total designed sample.}

\item{seed}{integer value used to set the state of the random number
generator (optional).}

\item{return_fit}{Returns summary if FALSE (default), otherwise return cmdstanr fit}

\item{num_iter}{Number of post warmup iterations}

\item{chains}{Number of chains (will be run in parallel)}

\item{model}{One of "mlogit" (the default) or "logit"}
}
\value{
A list with model fit (if return_fit=TRUE) and a \code{tibble}
estimates including point estimates for each party (median)
and limits of credible intervals.
}
\description{
Compute point estimate and credible intervals for each candidate.
}
\details{
Posterior simulations of parameters are computed using stan,
and each party's votes are simulated for every polling station (logit model)
or with a softmax link for the default (mlogit model). There is one
independent model for each party, and proportions are calculated from posterior simulations
of total votes.
}
