% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrp_estimation.R
\name{mrp_estimation}
\alias{mrp_estimation}
\title{Compute all postratified estimates}
\usage{
mrp_estimation(data, ..., stratum, frac = 1, n_iter = 2000,
  n_burnin = 500, n_chains = 3, seed = NA, parallel = FALSE,
  mc_cores = 6, model_string = NULL, set_strata_na = integer(0))
}
\arguments{
\item{data}{A \code{data.frame} with variables: ln_total, region,
distrito_loc_17, tamano_md, tamano_gd, casilla_ex and the column with
number of votes for the party.}

\item{...}{One or more unquoted expressions separated by commas, indicating
the column names with the votes for each candidate.}

\item{stratum}{If sampling the data, unquoted variable indicating the column
from the data.frame to be used as strata. The strata will also be used in
the hierarchical structure of the model.}

\item{frac}{If sampling the data, numeric value indicating the fraction
of the data to sample, the sample is selected using stratified sampling
with probability proportional to size.}

\item{n_iter}{Number of iterations, burnin size and chains.
to be used in \code{\link[R2jags]{jags}}.}

\item{n_burnin}{Number of iterations, burnin size and chains.
to be used in \code{\link[R2jags]{jags}}.}

\item{n_chains}{Number of iterations, burnin size and chains.
to be used in \code{\link[R2jags]{jags}}.}

\item{seed}{Integer value used to set the state of the random number
generator.}

\item{parallel}{Logical value indicating whether to parallelize the models,
if TRUE package parallel must be installed (uses mclapply and can not be
used in Windows).}

\item{mc_cores}{If parallelizing, the number of cores to use, parameter is
used in \code{\link[parallel]{mclapply}}}

\item{model_string}{String indicating the model to be used, if NULL defaults
to \code{"model_bern_t"}, also available \code{"model_t"}.}

\item{set_strata_na}{Option to exclude strata when fitting the model, used
for model evaluation and calibration.}
}
\value{
A \code{list} with the object fitted using R2jags::jags and a
  data.frame with the estimation summary (posterior means, medians, standard
  deviations and probability intervals per party).
}
\description{
This function calls \code{\link{mrp_party_estimation}} using all the parties
specified in `...`, once there is an estimation of the number of votes for
each party/candidate it computes the proportion of votes for each one.
}
\examples{
data("gto_2012")
mrp_estimation(gto_2012, party = pri_pvem:otros, stratum = distrito_loc_17,
  frac = 0.01, seed = 2212)
}
