% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrp_estimation_stan.R
\name{mrp_estimation_stan}
\alias{mrp_estimation_stan}
\title{Postratified estimation for the number of votes for several candidates.}
\usage{
mrp_estimation_stan(data, stratum, n_iter = 500, n_warmup = 250,
  n_chains = 1, seed = NA, model_string = NULL,
  set_strata_na = integer(0), partidos = c("RAC", "JAMK", "AMLO", "JHRC",
  "OTROS"), frame_name = "marco_nal_2018")
}
\arguments{
\item{data}{A \code{data.frame} with variables: ln_total, region,
distrito_loc_17, tamano_md, tamano_gd, casilla_ex and the column with
number of votes for the party.}

\item{stratum}{If sampling the data, unquoted variable indicating the column
from the data.frame to be used as strata. The strata will also be used in
the hierarchical structure of the model.}

\item{n_iter}{Number of iterations, burnin size and chains.
to be used in \code{\link[R2jags]{jags}}.}

\item{n_warmup}{burnin size}

\item{n_chains}{Number of iterations, burnin size and chains.
to be used in \code{\link[R2jags]{jags}}.}

\item{seed}{Integer value used to set the state of the random number
generator.}

\item{model_string}{String indicating the model to be used, if NULL defaults
to \code{"model_bern_t"}, also available \code{"model_t"}.}

\item{set_strata_na}{Option to exclude strata when fitting the model, used
for model evaluation and calibration.}
}
\description{
The function fits a model using the \code{rstan} package and
predicts proportional of votes for all candidates candidate in unobserved polling
stations. Optionally the model can be fit with a stratified random sample
}
