% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrp_party_estimation.R
\name{mrp_party_estimation}
\alias{mrp_party_estimation}
\title{Postratified estimation for the number of votes for a given candidate.}
\usage{
mrp_party_estimation(data, party, stratum, frac = 1, n_iter = 1000,
  n_burnin = 500, n_chains = 2, seed = NA, seed_jags = NA,
  model_string = NULL, set_strata_na = integer(0))
}
\arguments{
\item{data}{A \code{data.frame} with variables: ln_total, region,
distrito_loc_17, tamano_md, tamano_gd, casilla_ex and the column with
number of votes for the party.}

\item{party}{Unquoted variable indicating the column from the data.frame to
be modeled.}

\item{stratum}{If sampling the data, unquoted variable indicating the column
from the data.frame to be used as strata. The strata will also be used in
the hierarchical structure of the model.}

\item{frac}{If sampling the data, numeric value indicating the fraction
of the data to sample, the sample is selected using stratified sampling
with probability proportional to size.}

\item{n_iter, n_burnin, n_chains}{Number of iterations, burnin size and chains.
to be used in \code{\link[R2jags]{jags}}.}

\item{seed}{Integer value used to set the state of the random number
generator.}

\item{seed_jags}{Seed for the call \code{\link[R2jags]{jags}}.}

\item{model_string}{String indicating the model to be used, if NULL defaults
to \code{"model_bern_t"}, also available \code{"model_t"}.}
}
\value{
A \code{list} with the object fitted using R2jags::jags and the vector
  of simulated counts per candidate.
}
\description{
The function fits a model using the \code{\link[R2jags]{jags}} function and
predicts number of votes for a given candidate in unobserved polling
stations. Optionally the model can be fit with a stratified random sample
of the data, and predict for the population.
}
\examples{
# predict number of votes for pan_na using 6\% of the sample
mrp_gto_pan <- mrp_party_estimation(gto_2012, pan_na,
    stratum = distrito_loc_17, frac = 0.06, n_iter = 200, n_burnin = 100,
    n_chains = 2, seed = 19291)
mrp_gto_pan$fit
}
\seealso{
\code{\link{mrp_estimation}}
}
