% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnp_estimation.R
\name{nnp_estimation}
\alias{nnp_estimation}
\title{Normal No-Pooling model}
\usage{
nnp_estimation(data, ln, stratum, data_stratum, n_stratum, ...,
  n_sims = 1000)
}
\arguments{
\item{data}{\code{data.frame}}

\item{ln}{Unquoted variable indicating the nominal list (number of potential 
voters) at each polling station.}

\item{stratum}{Unquoted variable indicating the stratum for each polling
station.}

\item{data_stratum}{Data frame with stratum variable (named exactly as in 
\code{data}) and number of polling stations per strata.}

\item{n_stratum}{Unquoted variable indicating the number of polling stations
in each stratum.}

\item{...}{Unquoted variables indicating the number of votes in each polling
station for each candidate.}

\item{std_errors}{Logical value indicating whether to compute standard errors
(using bootstrap), defaults to TRUE.}

\item{B}{Number of bootstrap replicates used to compute standard errors, 
defaults to 50.}

\item{seed}{integer value used to set the state of the random number
generator (optional). It will only be used when computing standard errors.}
}
\value{
A list with two entries: 
\enumerate{
  \item lambdas_summary \code{data.frame} including posterior mean, median, 
  standard error, and quantiles (0.025 and 0.975) for each party. 
  \item lamdas_sim \code{data.frame} with simulations for each party.
}
}
\description{
This function implements the Normal No-Pooling model from \cite{AA}.
}
\examples{
# count number of polling stations per stratum
gto_stratum_sizes <- gto_2012 \%>\%
    dplyr::group_by(distrito_loc_17) \%>\%
    dplyr::summarise(n_stratum = n())
gto_sample <- select_sample_prop(gto_2012, stratum = distrito_loc_17, 0.06)
gto_nnp <- nnp_estimation(data = gto_sample, ln = ln_total, 
  stratum = distrito_loc_17, 
  data_stratum = gto_stratum_sizes, n_stratum = n_stratum, 
  pri_pvem:otros, n_sims = 100)
gto_nnp$lambdas_summary
}
