% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_election_day.R,
%   R/process_election_day_ratio.R
\name{process_batch}
\alias{process_batch}
\alias{ratio_process_batch}
\title{Automatically process batch of new data, and write estimates in correct
form for INE systems}
\usage{
process_batch(
  path_name,
  file_name,
  log_file,
  path_out,
  path_mailbox,
  team = "default",
  even = "0",
  n_iter = 300,
  n_chains = 4,
  n_warmup = 200,
  adapt_delta = 0.8,
  max_treedepth = 10,
  nominal_max = 1200,
  seed = 221285,
  inv_metric_file = NULL
)

ratio_process_batch(path_name, file_name, path_out, B, team = "default")
}
\arguments{
\item{path_name}{Path to a file that will be used for estimation. On election
day it will be a file with a subset of the sample.}

\item{file_name}{Name of the file with the data.}

\item{log_file}{Path to logfile of process}

\item{path_out}{Path to directory where partial results will be
saved.}

\item{path_mailbox}{path to mailbox}

\item{team}{Name of team running the model, to be used in INE reports.}

\item{even}{skipping of batches}

\item{n_iter}{number of stan sampling iterations}

\item{n_chains}{number of stan chains}

\item{n_warmup}{numer of stan warmup iterations}

\item{adapt_delta}{The adaptation target acceptance statistic (default 0.80.}

\item{max_treedepth}{The maximum allowed tree depth for the NUTS engine (default 10)}

\item{nominal_max}{maximum number of votes in special stations}

\item{seed}{random seed}

\item{inv_metric_file}{path to file with initial metric}

\item{B}{Number of bootstrap replicates used to compute standard errors,
defaults to 50.}
}
\description{
Automatically process batch of new data, and write estimates in correct
form for INE systems

Automatically process batch of new data, and write estimates in correct
form for INE systems
}
