% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_batch_election_day.R
\name{process_batch_election_day}
\alias{process_batch_election_day}
\alias{process_batch}
\alias{process_batch_stan}
\title{Functions for estimating on election day}
\usage{
process_batch(path_name, file_name, path_out, path_results, team = "default",
  n_iter = 1500, n_burnin = 500, n_chains = 1, parallel = TRUE)

process_batch_stan(path_name, file_name, path_out, path_results,
  team = "default", n_iter = 1500, n_warmup = 250, n_chains = 1)
}
\arguments{
\item{path_name}{Path to a file that will be used for estimation. On election
day it will be a file with a subset of the sample.}

\item{file_name}{Name of the file with the data.}

\item{path_out}{Path to directory where diagnostics, partial results, and 
data will be saved. Data is included in case there is need of additional 
checks.}

\item{path_results}{Path to the directory where partial results will be 
saved.}

\item{team}{Name of team running the model, to be used in INE reports.}

\item{n_iter}{Number of iterations, burnin size and chains.
to be used in \code{\link[R2jags]{jags}}.}

\item{n_burnin}{Number of iterations, burnin size and chains.
to be used in \code{\link[R2jags]{jags}}.}

\item{n_chains}{Number of iterations, burnin size and chains.
to be used in \code{\link[R2jags]{jags}}.}

\item{parallel}{Logical value indicating whether to parallelize the models,
if TRUE package parallel must be installed (uses mclapply and can not be
used in Windows).}
}
\description{
These functions are to be used on election day (July 1st, 2018), for 
estimating national election results and Governor election results in the 
States of Chiapas, Guanajuato and Morelos. The functions can be used within 
an R session but were created to be called autamtically by a Python script. 
The output was creted to feed INE's system.
}
\seealso{
\code{\link{mrp_estimation}}, \code{\link{mrp_party_estimation}}, 
  \code{\link{marco_2018}}
}
