% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_estimation.R
\name{ratio_estimation}
\alias{ratio_estimation}
\title{Ratio estimator to compute proportion of votes allocated to each party}
\usage{
ratio_estimation(data, stratum, n_stratum, std_errors = TRUE, B = 50,
  seed = NA, ...)
}
\arguments{
\item{data}{\code{data.frame}}

\item{n_stratum}{unquoted variable indicating the number of polling stations
in each stratum.}

\item{std_errors}{binary value indicating whether to compute standard errors
(bootstrap).}

\item{B}{number of bootstrap replicates used to compute standard errors.}

\item{seed}{integer value used to set the state of the random number
generator (optional). It will only be used when computing standard errors.}

\item{...}{unquoted variables indicating the number of votes in each polling
station for each candidate.}
}
\value{
A \code{list} with the object fitted using R2jags::jags and the
vector of simulated counts per candidate.
}
\description{
Compute ratio estimator for each candidate, standard errors are computed
with bootstrap resampling within each stratum and computing the standard
error of the samples (no corrections).
}
\details{
The bootstrap approach we use is not suitable
when the number of sampled polling stations within a strata is small.
Coverage might improve if confidence intervals are constructed with ABCs or
t-tables.
}
\examples{
# count number of polling stations per stratum
gto_stratum_sizes <- gto_2012 \%>\%
    dplyr::group_by(distrito_loc_17) \%>\%
    dplyr::mutate(n_stratum = n())
# stratified random sample (size 6\%), sample size proportional to strata size
gto_sample <- select_sample_prop(gto_stratum_sizes,
stratum = distrito_loc_17, 0.06)
gto_sample \%>\%
    ratio_estimation(stratum = distrito_loc_17, n_stratum = n_stratum,
    ... = pri_pvem:otros)
}
